Refactor command groups
