+++ ecstools/main.py
@@ -2,40 +2,20 @@ import os
 import sys
 import boto3
 import click
-import string
 
 from botocore.exceptions import ProfileNotFound, NoRegionError
-
+from ecstools.lib.cli import MyCLI
 
 version = '0.1.3'
 
-plugin_folder = os.path.join(os.path.dirname(__file__), 'commands')
+commands_dir = os.path.join(os.path.dirname(__file__), 'commands')
 
 
-class MyCLI(click.MultiCommand):
-
-    def list_commands(self, ctx):
-        rv = []
-        alpha = string.ascii_letters
-        for filename in os.listdir(plugin_folder):
-            if filename.startswith(tuple(alpha)) and filename.endswith('.py'):
-                rv.append(filename[:-3])
-        rv.sort()
-        return rv
-
-    def get_command(self, ctx, name):
-        ns = {}
-        fn = os.path.join(plugin_folder, name + '.py')
-        if not os.path.isfile(fn):
-            click.echo('Command not found')
-            sys.exit(0)
-        with open(fn) as f:
-            code = compile(f.read(), fn, 'exec')
-            eval(code, ns, ns)
-        return ns['cli']
+class Commands(MyCLI):
+    plugin_folder = commands_dir
 
 
-@click.group(cls=MyCLI)
+@click.group(cls=Commands)
 @click.pass_context
 @click.version_option(version=version, message=version)
 @click.option('-p', '--profile', help='AWS profile')
